refactor: split `soil_id/graphql/soil_data.py`
